Added some validation to ensure the application does not crash. For Example: Clicking Convert & Save with an invalid/blank path caused the application to crash.

+++ Wirenut Compiler/Form1.vb
@@ -18,47 +18,56 @@ Public Class Form1
     End Sub
 
     Private Sub convertSave_Click(sender As Object, e As EventArgs) Handles convertSave.Click
+
         Dim openPath As String = folderOpenPath.Text
         Dim savePath As String = folderSavePath.Text
 
-        Dim modFiles As String() = Directory.GetFiles(openPath, "*.jar")
-
-        If (Not Directory.Exists(openPath + "/mods")) Then
-
-            Directory.CreateDirectory(openPath + "/mods")
+        If (openPath IsNot "" And savePath IsNot "" And Directory.Exists(openPath) And Directory.Exists(savePath)) Then
 
 
-        End If
+            Dim modFiles As String() = Directory.GetFiles(openPath, "*.jar")
 
-        If (Not Directory.Exists(openPath + "/done")) Then
+            If (Not Directory.Exists(openPath + "/mods")) Then
 
-            Directory.CreateDirectory(openPath + "/done")
-
-
-        End If
-
-        progressBar.Maximum = modFiles.Count
+                Directory.CreateDirectory(openPath + "/mods")
 
 
-        For Each modFile As String In modFiles
-            Dim fileName As String = Path.GetFileNameWithoutExtension(modFile)
-            Dim fileExtension As String = Path.GetExtension(modFile)
-
-
-            System.IO.File.Move(modFile, openPath + "/mods/" + fileName + fileExtension)
-            Console.WriteLine("Moved To Mods Folder")
-
-            If (Not File.Exists(savePath + "/" + fileName.ToLower + ".zip")) Then
-
-                ZipFile.CreateFromDirectory(openPath + "/mods", savePath + "/" + fileName.ToLower + ".zip", CompressionLevel.Fastest, True)
-                Console.WriteLine("Zipped File")
             End If
-            System.IO.File.Move(openPath + "/mods/" + fileName + fileExtension, openPath + "/done/" + fileName + fileExtension)
-            Console.WriteLine("Moved To Done Folder")
 
-            progressBar.Value = progressBar.Value + 1
-            Console.WriteLine("Updated Progress Bar")
-        Next
+            If (Not Directory.Exists(openPath + "/done")) Then
+
+                Directory.CreateDirectory(openPath + "/done")
+
+
+            End If
+
+            progressBar.Maximum = modFiles.Count
+
+
+            For Each modFile As String In modFiles
+                Dim fileName As String = Path.GetFileNameWithoutExtension(modFile)
+                Dim fileExtension As String = Path.GetExtension(modFile)
+
+
+                System.IO.File.Move(modFile, openPath + "/mods/" + fileName + fileExtension)
+                Console.WriteLine("Moved To Mods Folder")
+
+                If (Not File.Exists(savePath + "/" + fileName.ToLower + ".zip")) Then
+
+                    ZipFile.CreateFromDirectory(openPath + "/mods", savePath + "/" + fileName.ToLower + ".zip", CompressionLevel.Fastest, True)
+                    Console.WriteLine("Zipped File")
+                End If
+                System.IO.File.Move(openPath + "/mods/" + fileName + fileExtension, openPath + "/done/" + fileName + fileExtension)
+                Console.WriteLine("Moved To Done Folder")
+
+                progressBar.Value = progressBar.Value + 1
+                Console.WriteLine("Updated Progress Bar")
+            Next
+        Else
+
+            MessageBox.Show("Please choose a Valid Folder location!")
+
+        End If
 
     End Sub
 End Class
